requirements_dev.txt
alembic==1.4.1
apache-airflow==2.0.2
apache-airflow-providers-ftp==1.1.0
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-mysql==1.1.0
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
bidict==0.21.2
biosppy==0.7.0
blinker==1.4
cached-property==1.5.2
cattrs==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
cloudpickle==1.6.0
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
connexion==2.7.0
coverage==5.5
croniter==0.3.37
cryptography==3.4.7
cycler==0.10.0
databricks-cli==0.14.3
dataclasses==0.7
defusedxml==0.7.1
dill==0.3.3
dnspython==2.1.0
docker==5.0.0
docutils==0.17.1
ecg-qc==1.0b3
email-validator==1.1.2
entrypoints==0.3
flake8==3.9.2
Flask==1.1.4
Flask-AppBuilder==3.3.0
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gatspy==0.3
gitdb==4.0.7
GitPython==3.1.17
graphviz==0.16
gunicorn==19.10.0
h5py==3.1.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iniconfig==1.1.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.4
mccabe==0.6.1
mlflow==1.17.0
mysql==0.0.3
mysql-connector-python==8.0.22
mysqlclient==2.0.3
natsort==7.1.1
numpy==1.19.5
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==20.9
pandas==1.1.5
pathtools==0.1.2
pendulum==2.1.2
pep562==1.1
Pillow==8.2.0
pluggy==0.13.1
prison==0.1.3
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
protobuf==3.17.1
psutil==5.8.0
psycopg2==2.8.6
py==1.10.0
py-ecg-detectors==1.0.2
pycodestyle==2.7.0
pycparser==2.20
pyEDFlib==0.1.22
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-cov==2.10.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
rich==9.2.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
setproctitle==1.2.2
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.4
sqlparse==0.4.1
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
typing==3.7.4.3
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.4
websocket-client==1.0.1
Werkzeug==1.0.1
wfdb==3.3.0
WTForms==2.3.3
xgboost==1.4.2
zipp==3.4.1
